// packages/langium-crmscript/src/generated/ast.ts
export type Expression = BinaryExpression | BooleanExpression | MemberCall | NilExpression | NumberExpression | StringExpression | UnaryExpression;

export const Expression = {
    $type: 'Expression'
} as const;

export function isExpression(item: unknown): item is Expression {
    return reflection.isInstance(item, Expression.$type);
}

export interface ExpressionBlock extends langium.AstNode {
    readonly $container: ExpressionBlock | ForStatement | FunctionDeclaration | IfStatement | ImplementationModel | WhileStatement;
    readonly $type: 'ExpressionBlock';
    statements: Array<Statement>;
    types: Array<Type>;
}

export const ExpressionBlock = {
    $type: 'ExpressionBlock',
    statements: 'statements',
    types: 'types'
} as const;
// ...
export interface ForStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'ForStatement';
    block: ExpressionBlock;
    condition?: Expression;
    counter?: NamedElement;
    execution?: ForExecution;
}

export const ForStatement = {
    $type: 'ForStatement',
    block: 'block',
    condition: 'condition',
    counter: 'counter',
    execution: 'execution'
} as const;

export function isForStatement(item: unknown): item is ForStatement {
    return reflection.isInstance(item, ForStatement.$type);
}

export interface FunctionDeclaration extends langium.AstNode {
    readonly $container: DefinitionModel | ExpressionBlock | ForStatement | ImplementationModel;
    readonly $type: 'FunctionDeclaration';
    array: boolean;
    arrayOfArray: boolean;
    body: ExpressionBlock;
    name: string;
    parameters: Array<Parameter>;
    returnType: langium.Reference<Class>;
}

export const FunctionDeclaration = {
    $type: 'FunctionDeclaration',
    array: 'array',
    arrayOfArray: 'arrayOfArray',
    body: 'body',
    name: 'name',
    parameters: 'parameters',
    returnType: 'returnType'
} as const;
// ...
export interface IfStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'IfStatement';
    block: ExpressionBlock;
    condition: Expression;
    elseBlock?: ExpressionBlock;
}

export const IfStatement = {
    $type: 'IfStatement',
    block: 'block',
    condition: 'condition',
    elseBlock: 'elseBlock'
} as const;

export function isIfStatement(item: unknown): item is IfStatement {
    return reflection.isInstance(item, IfStatement.$type);
}

export interface ImplementationModel extends langium.AstNode {
    readonly $type: 'ImplementationModel';
    includes: Array<Include>;
    statements: Array<Statement>;
    types: Array<Type>;
}

export const ImplementationModel = {
    $type: 'ImplementationModel',
    includes: 'includes',
    statements: 'statements',
    types: 'types'
} as const;

export function isImplementationModel(item: unknown): item is ImplementationModel {
    return reflection.isInstance(item, ImplementationModel.$type);
}

export interface Include extends langium.AstNode {
    readonly $container: ImplementationModel;
    readonly $type: 'Include';
    file: string;
}

export const Include = {
    $type: 'Include',
    file: 'file'
} as const;
// ...
export type NamedElement = Class | Enum | EnumMember | FieldMember | FunctionDeclaration | MethodMember | VariableDeclaration;

export const NamedElement = {
    $type: 'NamedElement'
} as const;
// ...
export interface PrintStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'PrintStatement';
    value: StringExpression;
}

export const PrintStatement = {
    $type: 'PrintStatement',
    value: 'value'
} as const;

export function isPrintStatement(item: unknown): item is PrintStatement {
    return reflection.isInstance(item, PrintStatement.$type);
}

export interface ReturnStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'ReturnStatement';
    value?: Expression;
}

export const ReturnStatement = {
    $type: 'ReturnStatement',
    value: 'value'
} as const;

export function isReturnStatement(item: unknown): item is ReturnStatement {
    return reflection.isInstance(item, ReturnStatement.$type);
}

export type Statement = Expression | ExpressionBlock | ForStatement | IfStatement | PrintStatement | ReturnStatement | WhileStatement;

export const Statement = {
    $type: 'Statement'
} as const;
// ...
export type Type = FunctionDeclaration | NamedElement;

export const Type = {
    $type: 'Type'
} as const;
// ...
export interface WhileStatement extends langium.AstNode {
    readonly $container: ExpressionBlock | ImplementationModel;
    readonly $type: 'WhileStatement';
    block: ExpressionBlock;
    condition: Expression;
}

export const WhileStatement = {
    $type: 'WhileStatement',
    block: 'block',
    condition: 'condition'
} as const;